Changes needed to OGR code for 64 bit IDs.
The OGR library doesn't support 64 bit integers. So we have to either use
doubles to represent large IDs or we stay with the 32 bit integers where we are
sure that the IDs are small enough, ie. for ways and relations.

+++ examples/osmium_toogr2.cpp
@@ -103,7 +103,8 @@ public:
                 OGRFeature* feature = OGRFeature::CreateFeature(m_layer_mp->GetLayerDefn());
                 OGRMultiPolygon* ogrmp = Osmium::Geometry::create_ogr_geometry(mp);
                 feature->SetGeometry(ogrmp);
-                feature->SetField("id", area->id());
+                // there are not so many multipolygon relations, so these ids will still fit into a 32bit int
+                feature->SetField("id", static_cast<int>(area->id()));
 
                 if (m_layer_mp->CreateFeature(feature) != OGRERR_NONE) {
                     std::cerr << "Failed to create feature.\n";
